+++ src/scripts/text-input/directives.js
@@ -1,0 +1,17 @@
+(function (angular, _, moment) {
+    'use strict';
+
+    angular.module('aif-text-input', []).directive('aifTextInput', [
+        function () {
+            return {
+                restrict: 'A',
+                replace: true,
+                scope: {
+                    model: '=',
+                    disable: '=?'
+                },
+                templateUrl: 'templates/aif/text-input/template.html'
+            };
+        }
+    ]);
+}(window.angular, window._, window.moment));